refactor and correct Client classes
